requirements.txt
requests
dotenv
ShopifyAPI
Pillow
openai